requirements.txt
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
openai>=1.0.0
python-dotenv>=1.0.0
sqlglot>=25.0.0
streamlit>=1.42.0
pandas>=2.2.0
